requirements.txt
langchain
langchain-ollama
faiss-cpu
sentence-transformers
scikit-learn
unstructured
python-dotenv
